Add MultiEWMA to make Meter updates atomic
We remove the lock in (*StandardMeter).Mark because each EWMAs shares
the same counter. Snapshots will use the MultiEWMA members instead.
The rates will be consistent with each other as rate updates are still
synchronised.

The only change in behaviour is that the increments of StandardMeter.count
and MultiEWMA.uncounted by (*StandardMeter).Mark are no longer atomic.
This only matters when ticking the meter every 5 seconds, where threads that
called Mark before the 5-second tick may appear to have called it after the
tick instead.